Complete move of filter related extensions out of Filter
